Delete RegisterProfile after activation, and some test refactoring.
